refactor(firmware): sensor frame is header, counter, payload, checksum
The frame layout is now spelled out as the specification gives it - two
header bytes, one counter byte, the payload, two checksum bytes - rather
than being implicit in the code. Only the payload size is a parameter.

It is not much of a parameter. Frame size times frame rate has to equal the
line rate, and 180 Hz on a 115200 baud line leaves exactly 64 bytes per
frame, so the payload is 59. Building with a 64 byte payload makes the frame
69 bytes and fails three assertions at once: the line rate no longer
matches, the sensor no longer yields exactly 16 frames per second, and a
1024 byte burst is no longer a whole number of wire units. A 69 byte frame
at 180 Hz would need 12420 B/s on a line that carries 11520 and would
overflow the 4096 byte ring in 4.5 seconds.

The counter is one byte now instead of two, so it wraps every 256 frames -
1.42 seconds. Gap detection uses unsigned 8 bit arithmetic and stays correct
across the wrap; the only loss it cannot see is exactly 256 consecutive
frames.

Re-measured after the change: 164.3 Hz under CAN load, 6468 CAN frames in
44 bursts, 704 units reassembled, every error counter still zero.

+++ firmware/src/engine/engine.c
@@ -252,7 +252,7 @@ void engine_protocol_task(void *arg)
 void engine_sensor_task(void *arg)
 {
     (void)arg;
-    uint16_t counter = 0;
+    uint8_t  counter = 0;
     uint8_t  frame[FRAME_LEN];
 
     /* Prime the ring before the line starts, so the first byte tick has
